Isolated the 180 scan into its own files scanner.h and scanner.c for ease of use in the future
