+++ Main Code/Lab 2/Debug/ccsObjs.opt
@@ -1,1 +1,1 @@
-"./adc.obj" "./cliffSensor.obj" "./init.obj" "./lcd.obj" "./main.obj" "./movement.obj" "./movementTest.obj" "./open_interface.obj" "./servo.obj" "./sonar.obj" "./timer.obj" "./tm4c123gh6pm_startup_ccs.obj" "./uart.obj" "./wifi.obj" "../tm4c123gh6pm.cmd" -llibc.a -l"C:/ti/TivaWare_C_Series-2.1.2.111/driverlib/ccs/Debug/driverlib.lib" 
+"./adc.obj" "./cliffSensor.obj" "./init.obj" "./lcd.obj" "./main.obj" "./movement.obj" "./movementTest.obj" "./open_interface.obj" "./scanner.obj" "./servo.obj" "./sonar.obj" "./timer.obj" "./tm4c123gh6pm_startup_ccs.obj" "./uart.obj" "./wifi.obj" "../tm4c123gh6pm.cmd" -llibc.a -l"C:/ti/TivaWare_C_Series-2.1.2.111/driverlib/ccs/Debug/driverlib.lib"
