+++ Main Code/Lab 2/Debug/main2.d
@@ -1,0 +1,103 @@
+# FIXED
+
+main2.obj: ../main2.c
+main2.obj: ../lcd.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/linkage.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdarg.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdint.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h
+main2.obj: C:/ti/TivaWare_C_Series-2.1.2.111/inc/tm4c123gh6pm.h
+main2.obj: ../timer.h
+main2.obj: ../movement.h
+main2.obj: ../open_interface.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdlib.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/math.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/_defs.h
+main2.obj: ../Timer.h
+main2.obj: ../lcd.h
+main2.obj: ../cliffSensor.h
+main2.obj: ../init.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h
+main2.obj: ../sonar.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h
+main2.obj: ../timer.h
+main2.obj: ../lcd.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdbool.h
+main2.obj: C:/ti/TivaWare_C_Series-2.1.2.111/driverlib/interrupt.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/math.h
+main2.obj: ../uart.h
+main2.obj: ../timer.h
+main2.obj: ../wifi.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h
+main2.obj: ../uart.h
+main2.obj: ../timer.h
+main2.obj: ../adc.h
+main2.obj: ../lcd.h
+main2.obj: ../servo.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h
+main2.obj: ../timer.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/math.h
+main2.obj: U:/git/PrawnStars/Main\ Code/Lab\ 2/scanner.h
+main2.obj: U:/git/PrawnStars/Main\ Code/Lab\ 2/init.h
+main2.obj: U:/git/PrawnStars/Main\ Code/Lab\ 2/uart.h
+main2.obj: U:/git/PrawnStars/Main\ Code/Lab\ 2/timer.h
+main2.obj: U:/git/PrawnStars/Main\ Code/Lab\ 2/servo.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h
+main2.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/math.h
+
+../main2.c: 
+../lcd.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/linkage.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdarg.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdint.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h: 
+C:/ti/TivaWare_C_Series-2.1.2.111/inc/tm4c123gh6pm.h: 
+../timer.h: 
+../movement.h: 
+../open_interface.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdlib.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/math.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/_defs.h: 
+../Timer.h: 
+../lcd.h: 
+../cliffSensor.h: 
+../init.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h: 
+../sonar.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h: 
+../timer.h: 
+../lcd.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdbool.h: 
+C:/ti/TivaWare_C_Series-2.1.2.111/driverlib/interrupt.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/math.h: 
+../uart.h: 
+../timer.h: 
+../wifi.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h: 
+../uart.h: 
+../timer.h: 
+../adc.h: 
+../lcd.h: 
+../servo.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h: 
+../timer.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/math.h: 
+U:/git/PrawnStars/Main\ Code/Lab\ 2/scanner.h: 
+U:/git/PrawnStars/Main\ Code/Lab\ 2/init.h: 
+U:/git/PrawnStars/Main\ Code/Lab\ 2/uart.h: 
+U:/git/PrawnStars/Main\ Code/Lab\ 2/timer.h: 
+U:/git/PrawnStars/Main\ Code/Lab\ 2/servo.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/string.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdio.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/math.h:
