+++ Main Code/Lab 2/Debug/tm4c123gh6pm_startup_ccs.d
@@ -1,7 +1,7 @@
 # FIXED
 
 tm4c123gh6pm_startup_ccs.obj: ../tm4c123gh6pm_startup_ccs.c
-tm4c123gh6pm_startup_ccs.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.1.LTS/include/stdint.h
+tm4c123gh6pm_startup_ccs.obj: C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdint.h
 
 ../tm4c123gh6pm_startup_ccs.c: 
-C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.1.LTS/include/stdint.h: 
+C:/ti/ccsv7/tools/compiler/ti-cgt-arm_16.9.4.LTS/include/stdint.h:
